put all type and intrf declarations in types file
Jira ticket: CAMS-288
