+++ user-interface/src/staff-assignment/filters/staffAssignmentFilter.types.ts
@@ -1,4 +1,6 @@
 import { ComboOption } from '@/lib/components/combobox/ComboBox';
+import { ResponseBody } from '@common/api/response';
+import { UstpOfficeDetails } from '@common/cams/offices';
 import { CamsUserReference } from '@common/cams/users';
 
 interface Store {
@@ -8,6 +10,10 @@ interface Store {
   setOfficeAssigneesError(val: boolean): void;
 }
 
+type StaffAssignmentFilterViewProps = {
+  viewModel: ViewModel;
+};
+
 interface ViewModel {
   officeAssignees: CamsUserReference[];
   officeAssigneesError: boolean;
@@ -16,4 +22,33 @@ interface ViewModel {
   handleFilterAssignee(assignees: ComboOption[]): void;
 }
 
-export type { Store, ViewModel };
+interface StaffAssignmentFilterRef {
+  refresh: () => void;
+}
+
+type StaffAssignmentScreenFilter = {
+  assignee: CamsUserReference;
+};
+
+type StaffAssignmentFilterProps = {
+  handleFilterAssignee(assignees: ComboOption[]): void;
+};
+
+interface StaffAssignmentFilterUseCase {
+  assigneesToComboOptions(officeAssignees: CamsUserReference[]): ComboOption[];
+  fetchAssignees(): void;
+  getOfficeAssignees(
+    apiFunction: (office: string) => Promise<ResponseBody<CamsUserReference[]>>,
+    offices: UstpOfficeDetails[],
+  ): Promise<CamsUserReference[]>;
+}
+
+export type {
+  Store,
+  ViewModel,
+  StaffAssignmentFilterViewProps,
+  StaffAssignmentFilterRef,
+  StaffAssignmentScreenFilter,
+  StaffAssignmentFilterProps,
+  StaffAssignmentFilterUseCase,
+};
